Fix warnings for the sparse methods.

+++ include/tatami/sparse/primary_extraction.hpp
@@ -27,7 +27,7 @@ std::pair<size_t, size_t> extract_primary_dimension(
     bool do_cache = !cached.empty();
     if (do_cache) {
         auto val = cached[i];
-        if (val.first != -1) {
+        if (val.first != static_cast<size_t>(-1)) {
             return val;
         }
     }
@@ -125,7 +125,7 @@ void primary_dimension(
     if (indices[0]) { // Only jumping ahead if the start is non-zero.
         bool do_cache = !cached.empty();
         if (do_cache) {
-            if (cached[i] != -1) { // retrieving the jump from cache, if we came here before.
+            if (cached[i] != static_cast<size_t>(-1)) { // retrieving the jump from cache, if we came here before.
                 iIt += cached[i];
             } else {
                 auto iIt2 = std::lower_bound(iIt, eIt, subset[0]);
